fix(codegen): move auth to standalone plugin
